+++ apps/client/app/(afterLogin)/my-study/page.tsx
@@ -17,6 +17,7 @@ const MyStudyPage = () => {
       <StudyCurriculum />
       <Space height={64} />
       <StudyNotices />
+      <Space height={35} />
     </>
   );
 };
