+++ apps/client/app/(beforeLogin)/(auth-error)/auth-error-after-recruitment/page.tsx
@@ -1,19 +1,11 @@
-"use client";
-
 import { css } from "@styled-system/css";
 import { Text } from "@wow-class/ui";
 import { routePath } from "constants/routePath";
 import Image from "next/image";
-import { useRouter } from "next/navigation";
+import Link from "next/link";
 import Button from "wowds-ui/Button";
 
 const AuthErrorAfterRecruitmentPage = () => {
-  const router = useRouter();
-
-  const handleClickNavigateHome = () => {
-    router.push(routePath.landing);
-  };
-
   return (
     <main className={mainContentStyle}>
       <Image
@@ -32,7 +24,7 @@ const AuthErrorAfterRecruitmentPage = () => {
         모집 기간에 합류 후 이용해주세요.
       </p>
       <div className={buttonContainerStyle}>
-        <Button aria-label="홈으로 이동" onClick={handleClickNavigateHome}>
+        <Button aria-label="홈으로 이동" asProp={Link} href={routePath.landing}>
           홈으로 이동
         </Button>
       </div>
