+++ package.json
@@ -46,6 +46,6 @@
     "clsx": "^2.1.1",
     "wowds-icons": "^0.1.3",
     "wowds-tokens": "^0.1.1",
-    "wowds-ui": "^0.1.8"
+    "wowds-ui": "^0.1.14"
   }
 }
